feat: 크레딧 결제 api 구현

+++ chaekit-spring/src/main/java/qwerty/chaekit/dto/ebook/credit/CreditWalletResponse.java
@@ -5,5 +5,5 @@ import lombok.Builder;
 @Builder
 public record CreditWalletResponse(
         Long walletId,
-        Integer balance
+        Long balance
 ) { }
